[http-client] Restore CA discovery logic when configuring SSL client.
This resolves a runtime bug triggered when `"/"` is passed into an
unsafe OpenSSL C block which is expecting a path to a CA certs file.
Instead, we continue to use the original fallback logic of looking at
environment variables, a `core/cacerts` package and an inlined copy to
ensure that the Hyper client is properly wired no matter what the
operating system.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>