Reduce allocations when working with ServiceGroup
ServiceGroup now contains a single string allocation and three
functions to access parts of the service group. This allows us
to greatly reduce the amount of string allocations when attempting
to comopare the full representation of a service group instead of
it's components.

I also did a bit of work refactoring the public API of the key
generation functions to make them easier to use

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>