Add Smoke Test hook and overall hook refactorings
This change adds a Smoke Test hook to the HookTable. It is currently
not executed because further refactorings are necessary to ensure
that hooks are run asynchronously without causing the main loop of
the manager to block.

This commit also contains a number of refactors to move hooks from
the package module in the supervisor crate to the service module.
While hook templates are defined in a package, they are largely a
component of the runtime and should be owned by the service module.

In a future commit we will hook up the smoke test and health check
hooks to background threads which will feed a cache on the service
struct to be read by the gossip server and http-gateway.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>