[Breaking sup] Remove initializer topology as leader can be used.
After several on-site customer engagements, the team has come to realize
that the initializer topology in most cases is a simplified version of
the leader topology. What has been missing to this point has been using
an application's booted "health" as a signal for other service group
members to proceed or stop booting their services (or during a software
update). The combination of leader topology and the rolling update
strategy should address the need for the initializer topology.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>