Refactor Hooks and add SmokeTest hook
This change stubs out a new hook to be run on service start after the
first initial successful health check. The Smoke Test hook is our first
long running hook which needs to be executed in a background thread
so I haven't hooked it up just yet. More refactoring is necessary to
the way we schedule a service to do work before the background runner
can be implemented.

In addition to the new hook this commit also refactors hooks into a
concrete struct implementing the new `Hook` trait eliminating the
HookType enum. This change allows us to move hook execution behaviour
into the hooks module. Running a hook now has an associated type for
it's return which can be massaged into any type. For example, the
`HealthCheckHook` and `SmokeTestHook` return a health/smoke check
instead of an i32. This will allow us to build more complex return
values from hooks and allows us to define them on the hook itself.

Lastly, hooks are no longer compiled on demand. This prior change
meant that people running a hook needed a mutable reference of the
Service struct. We now just compile after configuration and during
initialization.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>