Add `pkg_binds` and `pkg_binds_optional` plan variables
This adds two new plan variables:

* `pkg_binds`
* `pkg_binds_optional`

These plan variables define what is accepted by the supervisor's `--bind`
flag. `pkg_binds` defines what is *required* to be provided to the
`--bind` flag for the service to start.

This is an MVP of the consumer's portion of our inter-service contract.
The producer's side is defined by `pkg_exports`. In the future, this
metadata will also participate in the service's state machine to
determine when it is appropriate for a service to start up. They will
also prevent a service from binding to a group which does not export
the desired configuration.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>