Fix hook generation and reconfigure on config from gossip
In a previous refactor I had broken the timing on loading and
compiling hooks. This commit fixes those errors along with a
fix for reloading gossip configuration on config from gossip.
This was broken in a previous commit by an optimization to
how and when we load configuration from disk - we now explicitly
reload only the gossip portion of the service config on change.

Additionally, hook compilation will now happen on demand the first
time any hook is called since the last time the service config was
updated. This greatly simplifies timing around when to compile
hooks and also ensure we only compile hooks when we need to.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>