Fix hook generation and reconfigure on config from gossip
In a previous refactor I had broken the timing on loading and
compiling hooks. This commit fixes those errors along with a
fix for reloading gossip configuration on config from gossip.
This was broken in a previous commit by an optimization to
how and when we load configuration from disk - we now explicitly
reload only the gossip portion of the service config on change.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>