[sup,hab] Add back `hap-sup config` subcommand.
This change restores the Supervisor's `config` subcommand to preserve
the behavior of the Supervisor in constrained environments, such as
in containers.

Additionally, in the event that the desired package does not have a
`default.toml`, a message explaining this will be printed to the
standard error stream rather than the standard output stream to help
with programs or scripts which may consume the raw TOML data.

References #1689
Fixes #1729

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>