Skip auto-update on start for rolling update strategy
Prior to this commit the supervisor would check for an updated version
of the package being loaded as a service if any update strategy was
specified. This behaviour actually only makes sense for the "at-once"
strategy where services can update independent of each other and should
always be at the latest version of the specified package goal on start.

This commit ensures that service groups implementing the rolling
strategy don't have their packages updated on load and relies on the
coordination of the strategy to ensure the supervisor is running the
correct version of the package.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>