Upgrade to serde 0.9 and toml 0.3 APIs.
This change primarily upgrades the codebase from serde 0.8 to 0.9 and
consequently from toml 0.2 to 0.3. Other dependent crates were also
updated as a full update hasnt taken place in over a month.

Currently the urlencoded crate still takes a dependency on an older
bodyparser which in turn depends on serde 0.8 which is why there are 2
versions of these dependencies in the full graph.

Serde and toml Release Notes
----------------------------

* https://github.com/serde-rs/serde/releases/tag/v0.9.0
* https://github.com/alexcrichton/toml-rs/commit/f66d8bcf33530c858a502bfa170f2383a8cbc204

Of special note: the toml API updates in most cases introduce a
Result-wrapped return and therefore introduce more explicit cases where
the error may need to be handled. Some functions which use these API
calls don't return a Result themselves and therefore some uses of
`.expect()` were used (usually with a TODO left).

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>