Panic if run hook run() called directly
The run hook is a special hook and needs to be handled a bit differently
than the others. It's not really part of the lifecycle as much as it is
the main entry point for a service. Because of this and, unless the
supervisor needs to restart the process, it can't be run more than once
or out of turn in anyway. It also can't log it's output like another
lifecycle hook might.

This change overrides the run() function of the RunHook with a panic!
to warn the developer. In the future we should think about refactoring
the code a bit to be more clear about the special nature of the
run hook.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>