Use `as |thing|` form for all `#each` examples
Doing it this way removes confusing ambiguity.

If you're not using it and you want to refer to something with the same
key as the item you're iterating over, you have to do `../thing`, which
is an extra thing to learn and easily leads to errors and confusion.

Always using `as |thing|` makes it so you never have to think about
shadowing variables and you templates, while a smidge longer, are always
clear.

Signed-off-by: Nathan L Smith <smith@chef.io>