[plan-build.sh] Fix a `last_build.env` formatting regression.
This restores the format of the `./results/last_build.env` file's
`pkg_blake2bsum` line. Some speed improvements and refactoring in #1804
resulted in `hab pkg hash` to now output the checksum *and* filename
rather than simply the checksum. This fix only keeps the first token in
the command's result string--that is, the checksum only.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>