Implement sharding down to the data layer
This commit makes sharding work all the way down to the database level.
It brings changes to the core database code, and a refactor of how
connections are gotten from the shared pool.

Testing framework has changed; we now run a single migration on a
template database, and then create a copy for every test we run. The
results are very fast integration tests, without worrying about
re-running the entire sharded migration every time, and without special
case code to change sharding behavior at test time.

Additionally, we've added a functional test suite that starts at the API
and works backwards; it uses the same stack as our end to end tests for
the web ui.

The sessionsrv is migrated to PostgreSQL, and sharded on Account ID.

The jobsrv is migrated to use only the first shard, due to fairness.

The scheduler is migrated to use only the first shard.

The key work in originsrv is updated to route correctly.

Signed-off-by: Adam Jacob <adam@chef.io>