Add configuration option to disable SSL
NOTE: THIS IS FOR DEVELOPMENT PURPOSES ONLY!!

In production environments, we want to use SSL for the NGINX proxies
that sit in front of the API services. However, in development it may
be desireable to disable SSL for testing purposes, as we may not want
to generate and manage SSL certificates.

This commit adds a new `http.enable_ssl` option for the
builder-admin-proxy and builder-api-proxy configuration files. It
defaults to true.

Signed-off-by: Joshua Timberman <joshua@chef.io>