[sup] Implement `SpecWatcher` to pick up spec file changes on disk.
This change adds the `SpecWatcher` but doesn't wire it in anywhere yet
(it is been well tested in isolation in the meantime). The `SpecWatcher`
uses file system events (appropriate to the respective platforms) to
trigger a scan/load/reconcile activity which is the mechanism by which
the Supervisor's `Manager` will be able to add, remove, and (soon)
update running services.

Setting up the `SpecWatcher` is fairly trivial--invoke the `run()`
function giving it a directory path to watch (note this directory must
exist and is checked for):

```
let tmpdir = TempDir::new("specs").unwrap();
let mut watcher = SpecWatcher::run(tmpdir.path()).unwrap();
```

In order to furnish the `Manager` with an initial list of services to
load, the `SpecWatcher` provides a `.initial_events()` method. This
returns a `Vec<SpecWatcherEvent>` where each event is an
`AddService(ServiceSpec)` entry.

```
let initial_events = watcher.initial_events().unwrap();
```

Finally, on each `Manager` tick a `.new_events()` method is provided
which returns the same `Vec<SpecWatcherEvent>` but this time each event
could be an `AddService(ServiceSpec)` or a `RemoveService(ServiceSpec)`
entry. Think of this as tracking the state of a directory with version
control, each patch being a diff to the current desired state.

A few notes:

* While `SpecWatcher` deals with a changed `ServiceSpec` (i.e. where the
on-disk version for a service is different from what is currently
running), the implementation is the simple version: emit a
`RemoveService(..)` event followed by an `AddService(..)` event with the
new `ServiceSpec` desired state. This will, in effect, like restarting a
service. Not the end of the world but until (and if) we can support
updating a service in-place, this keeps us is a good(ish) place.
* The spec file names must conform to `<NAME>.spec.toml` where `<NAME>`
is the name component of a package identifier. In other words, a spec
file whos' contents are `ident = "core/redis"` must be named
`redis.spec.toml`. Failure to do so results in a warning message (with
hints as to why and how to fix it) and the file is explicitly skipped.
Remember that all spec files will be re-scanned many times in the future
so each scan is a chance that the user has fixed their problem and the
system will recover. In this way we try to fail-safe, or "do no further
harm".
* The same strategy is taken for malformed spec files or files that
cannot be properly read due to IO/permission issues/etc.: a warning
message is display and the file is explicitly skipped.
* Currently a file system event needs to fire for the scanning to occur.
There should probably be a timer that ensures the directory is being
scanned some minimum number of times per period (like at least once
every 30 seconds or 2 minutes). Most likely room for improvement
here--using this module live in the Supervisor will give us the feedback
we need.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>