[sup] Refactor `SpecWatcher` to call notify's sync API.
After putting `SpecWatcher` into use, I found the async APIs in notify
to not be as reliable as I expected (i.e. using `rx.try_recv`, etc. on
the channel), so this changes the strategy a bit. The new implementation
spawns a thread which sets up the channel and then consumes all fs
events with blocking reads. On each event it gets, it sets a shared
`AtomicBool` to true and the main checking function
`SpecWatcher.new_events()` simply checks the value of this shared
boolean, and resets it if it's loaded specs off disk. Much more
reliable!

References #1976

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>