[sup] Move location of adding/removing services in `Manager` tick loop.
This change enables the Supervisor to pick up new spec files in the spec
watch directory once the Supervisor is running.

This resolves a service starting issue once the Supervisor is
running. Relocating when we look for new services to add or old services
to remove allows `build_census()` to run before delegating to each
service's internal tick loop.

It remains to be seen if this is the correct location to remove services
(on first glance, I think so), but this will become clearer with more
exercise and testing.

References #1976

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>