Add mutually exclusive process lock to Supervisor
Since the Supervisor is writing and reading from disk, we can no
longer allow multiple Supervisors of the same "name" run at the
same time. This change prevents that from happening by writing a
LOCK file to the Supervisor's service directory.

Running two supervisors is not recommended, but it is supported by
means of the `--override-name` flag on the CLI. If this is not
specified, the Supervisor will now exit with a helpful error message
explaining how to start multiple.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>