Need a way to promote ChefConf
Adding a ChefConf link to the main navigation might not be the best way
to highlight the conference but it is a quick and easy way to link to
our conference without also taking over the home page's hero which is
currently being used for our 'next' Habitat event.

Open to other suggestions!

Signed-off-by: Nathen Harvey <nharvey@chef.io>