Add stop/load/unload/run commands to Supervisor
* `stop` - stop a running permanent or transient service. Stopped
	transient services will have their spec removed.
* `load` - write a service spec to be picked up by a running supervisor
	or a supervisor which will be started later. This will create
	a "persistent" service which will be started whenever the
	Supervisor is run
* `unload` - remove a service spec and stop the associated service if
	running
* `run` - start a supervisor with no services

This change also modifies the `start` command which now *requires* a
package identifier or file path to an artifact. Start will run a
Supervisor if one is not running. If one is already running, it will
add the given service as a transient service or start a stopped
persistent service.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>