Verify that specified bindings exist in census before service start.
When a service depends on a binding typically config files will be
rendered but in an unusable state. Plan authors must verify the presence
of the bindings settings in the init hook.
Delaying startup of the service will reduce output noise (from multiple
failed init hook runs) and cause less burdon on plan authors needing to
verify the config files consistency.

Signed-off-by: Justin Carter <justin@starkandwayne.com>