Add departed state for members
Fixes #1994.

This PR adds a fourth state for members, 'Departed'. A departed member
is no longer counted as valid for calculating the total population
needed for a quorum.

When a member is marked `Confirmed`, we record the time from the expired
thread. When the expired thread runs, it checks for members whose
departure time has come; and if it has, it sets their health to
`Departed`, removes them from the departure timer, and shares the
rumor.

This commit sets the departure timer at 72 hours, a particularly
conservative number. If an administrator does nothing to bring member
back to life in that timeframe, it will be marked departed
automatically.

Additionally, when a service group has confirmed members, and new
members are added to the service group, the supervisors will
automatically mark a confirmed member as departed. They do this in
alpha-sort order for the members ID's. This will ensure that we handle
the typical recovery scenario without any administrator intervention.
Hat tip to the clever folks at Consul for this solution.

You can manually depart a member with:

  `hab butterfly depart MEMBER_ID`

If a supervisor receives a rumor that it has been departed by an
administrator, it will exit with the signal to the launcher that it
should not be restarted. It also prints the reason why to the user (it
would be unsafe to assume you want the services to come back), with
guidance on how to remediate (delete the supervisors ring state).

Signed-off-by: Adam Jacob <adam@chef.io>