[sup] Restore support of the `--config-from` on `hab start`.
This is a bug fix change to restore the behavior of the `--config-from`
option for use in development. In a perverse twist, this behavior was
enforced by tests as the assumption at the time was this would only be
supported in the single-shot, development mode of running the
Supervisor. The end implementation however internally uses `ServiceSpec`
files consistently and as such we need to serialize (i.e. write) and
deserialize (i.e. read) this value. Restoring the (de)serializing logic
allows the user-provided option to flow through for the Supervisor to
use it. Sorry folks!

Fixes #2058

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>