[butterfly] Write dat random tmpfile locally then move.
This change fixes an issue when running the Supervisor on a Linux host
which has a separate `/tmp` filesystem, a `tmpfs` filesystem, etc. The
previous implmentation used `tempfile` crate which writes a file into
the system temporary file space such as `/tmp`, `/var/tmp`, etc. on a
Unix system. This means that if there is a separate filesystem, a file
move may cross a filesystem boundary which could fail, incur an IO
penalty if the file is large, etc. A better strategy is to make the file
in the same destination directory with a different, nondeterministic
name (i.e. random), then move it into place, becoming a more atomic
operation.

In my case, I am using an Arch Linux distro which is using zfs for the
userland and tmpfs for `/tmp`. As a result I was seeing many errors
containing the following:

```
Invalid cross-device link (os error 18)
```

Fixes #2078

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>