[plan-build] Allow Scaffolding packages to add build deps for Plans.
This change further refines the dependency resolving logic to allow a
Scaffolding package author to add **build** dependencies to a Plan. As a
result, the timing of the scaffolding loading is pulled before the full
build dependency resolving. This shouldn't affect any existing
Scaffolding packages as a Scaffolding package author should not be using
software dependencies that they have not directly added themselves. In
other words, if more build deps are required to run in the scaffolding
loading hook, then those deps should be added to the Scaffolding
package's `${pkg_deps[@]}`.

With this change, Scaffolding package can now hot-add more build
dependencies after inspecting the app's codebase.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>