Fix erroneous IO output when stopping a transient service
The manager handles cleaning up a service spec if the service is
stopped and it was a transient service. This commit makes a small
modification to the way we stop a service in the CLI. Instead of
removing the service spec if it is a transient service we instead
mark the service as stopped and allow the manager to handle the
details of how to go about doing it.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>