Remove a bad service spec on failure to load for Transient service
If a transient service is unable to start due to a service load
error we will now additionally remove the corresponding service
spec. Without this change, we would instead fail to load and
never remove the transient service spec since we only remove them
on service unload. Since we never successfully loaded, we never
actually unload and remove the spec

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>