[plan-build] Rename Scaffolding load function to `scaffolding_load()`.
This change updates the name of the optional load hook for a Scaffolding
package from the previously named `_scaffolding_begin()` to
`scaffolding_load()`. There are 2 reasons for the rename:

1. From the Scaffolding package author's point of view, this function is
part of the build system's public API and therefore should not be
considered an internal function. The build program itself will call this
function when appropriate, and this is internal implementation detail.
2. When explaining the Scaffolding functionality, it has been common to
say "when the Scaffolding is loaded". As such, a function name
containing the word `load` is probably more intuitive.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>