[plan-build] Introduce `do_before()` & `do_after()` build phases.
This change introduces 2 new build phases to the public build API:

* `do_before()` (with an empty `do_default_before()` implementation)
* `do_after()` (with an empty `do_default_after()` implementation)

After inspecting more Plans in the wild, it has become apparent that
there is a missing build phase which Plan authors would have used rather
than hooking into the `do_download()` phase to perform tasks like
dynamically calculating a new `$pkg_version` value, etc.

The `do_begin()` phase is often far too early in the build process to
perform many tasks as package dependencies are not yet loaded, the
`$PATH` is not yet set, and other build environment variables are not
yet present. The first publicly supported phase in which to perform
such tasks is `do_download()` which occurs directly after resolving
dependencies and setting up the build environment.

Rather than allowing this "hacked" pattern to continue unfettered, an
additional build phase is introduced **before** the downloading,
building, installing is started.

In an attempt to properly "bookend" the `do_before()` build phase, a
`do_after()` build phase is added after the `do_strip()` build phase. At
this point in the build, the package has been built, installed, and
stripped, but before the package metadata is written and before the
artifact is created and signed. This is a mirror to `do_begin()`
(extremely early) and `do_end()` (extremely late).

Future work may optionally exploit the `do_before()` phase by allowing a
documented location to recompute the `$pkg_version` value.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>