Logs!
Implements enough backend logic to create a "streaming build log"
experience, as well as laying the foundation for long-term log
storage.

Overview
========

Builder workers run `hab studio build` to create Habitat packages,
just as you would on your local workstation. Previously, the log
output of these jobs was written to disk on the worker machines, but
never sent anywhere else. We would like to a) stream this data to
users of our web UI so that they may monitor builds in real-time and
b) provide the ability to look at build output once a job has
completed.

To do this, several pieces of ZeroMQ messaging infrastructure are
added, allowing us to transmit log data from running jobs line-by-line
to the job server for consolidation and archiving, and for access via
the API. Both the worker server and the job server start log-related
processes in separate threads. The process on the worker server --
`LogForwarder` -- accepts lines of output from a build job (`LogPipe`)
and forwards them to the process on the job server -- `LogIngester` --
which accumulates the lines in files on disk, one per job. (It will
eventually also trigger archiving the logs in long-term storage
off-disk when they are complete, but that is not implemented here.)

We have separate threads and socket pairs handling the log processing
in order to decouple that processing from the actual execution of the
jobs, and to prevent log messages from filling up other more important
message queues (like actual job dispatch).

With log data centrally located on the job server, we can also create
an API for accessing a job's log data. Though we eventually would like
a Server-Sent Events (SSE) endpoint to stream log data to clients, the
current implementation does *not* use SSE (see below for why), but
uses polling instead (this is not long polling; it’s just making a
normal request periodically). The client makes an initial request and
receives back all the job's log data the server currently has, along
with offset data, as well as an indication of whether the log is
complete or not. If it is not complete, the client can wait and then
make another request, using the offset it received earlier. In this
way, the client can build up the entire log over a series of
requests. When the log is complete, the client can simply stop
polling.

Room for Improvement
====================

The polling API endpoint is the biggest target for
improvements. Because Iron and Hyper don’t currently support
asynchronous and non-blocking I/O, we can’t easily implement a SSE
endpoint for streaming logs back to clients. There are not really any
full-featured Rust SSE libraries yet. Websockets are more complexity
than this feature needs, and in any case, existing libraries appear to
all use a thread-per-connection approach. We can, however, write our
own small async server, based on something like
[Tokio](https://tokio.rs), and run it alongside our existing
Iron-based API server behind our Nginx proxy. While this won't
necessarily be difficult, we opted for the current polling approach to
more quickly implement the desired functionality and get feedback on
it sooner.

Log data “chunks” are currently sent from the workers one line at a
time. Once we have an SSE endpoint, this will give the most immediate
feedback to users, but we can increase the chunk size (e.g., multiple
lines at a time) we send out from workers if we encounter performance
issues.

When an API request for a portion of a job’s log comes in, we simply
read the file from disk and return the requested lines. Once we have
an SSE endpoint, the log chunk messages can simply be forwarded on to
the process serving those streams, thus eliminating this repeated file
reading.

Long-term, we’d like to ship completed logs off-disk to a remote
location like S3 (while allowing for S3-compatible solutions for
air-gapped or non-AWS installations) When a job log is finished, we
will send the logs off disk and record the URL at which they can be
found in the database entry for the job in a `log_url` field. The
presence of a value in this field indicates that a log is
“complete”. We don’t yet have the S3 upload functionality, so for now
we’ll simply add a `file://` URL (pointing to the on-disk location on
the job server) to the database upon completion.

Of Interest
===========

* The job server now listens on port 5569 for incoming log data from
  workers.

* The job server will not start if it cannot write job log data.

  Log data is accumulated on the job server in files-on-disk. The
  location is specifiable in config for the job server, defaulting to
  `/tmp`. The directory given must exist and be writable by the
  server, or the server will refuse to start.

* Workers now run with `HAB_NONINTERACTIVE` set.

  This disables the download progress bars in `hab studio build`
  output. Without this, we'd need to resort to hacks to deal with all
  the carriage returns and every "frame" of the "animation".

* Workers will refuse to start if multiple job servers exist

  Currently the Job Server is not sharded and we only run one. Thus,
  log chunks can be sent directly to the Job Server, instead of
  needing to be routed to the server that manages the particular job's
  shard (thus ensuring we consistently send logs to the same place for
  accumulation). If we were to shard the job server, the current
  implementation would fail horribly, with a single job's log output
  spread out across as many job servers as we have running. Thus, as a
  precaution, the LogMgr process will panic on startup if it is given
  multiple job servers to connect to, essentially preventing any
  workers from starting. This is more to guard against potential
  future breakage than anything.

* Logging uses a multi-frame ZMQ message protocol

  In order to notify clients that they can stop polling, and to know
  when to archive logs in long-term storage, we need to know when a
  log is complete. We use a two-frame message to convey this
  information. The first frame is a one-character code indicating the
  type of message that follows in the second frame. An `"L"` means a
  log chunk follows, while a `"C"` indicates a completion protocol
  message follows. The `LogIngester` takes action and parses messages
  based on this information.

Fixes #2024

Signed-off-by: Christopher Maier <cmaier@chef.io>