[plan-build] Add optional helper to dynamically compute pkg_version.
This change adds an alternative way to determine the version of a Plan
when building a package. In seeing more Plans in the wild, it has become
clear that some Plan authors wish to determine the package version by
cat'ing a common version file, using a Git command, or by some other
means which is dynamic.

There are 2 challenges to computing a package version:

* how the verison is computed
* when the version should be computed

To implement the "how", a Plan author needs to write a `pkg_version()`
function that prints the version string to standard output. The Plan
author then explicitly calls the new `update_pkg_version()` helper in a
build phase to determine the "when".

The `update_pkg_version()` can be called in the `do_before()` build
phase all the way up to but not including the `do_prepare()` build
phase, after which it is too late to safely recompute the other build
variables that depend on `$pkg_version`. If `update_pkg_version()` is
called too late, the build program will fail the build with an error
message.

Due to this new optional strategy, a Plan author can omit defining a
`pkg_version` variable, as long as they write a `pkg_version()`
function. This is checked for by the build program so that if the Plan
author omits the `pkg_version` variable and the `pkg_version()`
function, the build will fail.

Examples
--------

The following are partial Plan implementations with extra build phases
omitted for brevity.

This is a Plan which cats a static file in the source root of the
project to determine the version which can be safely determined in the
`do_before()` phase:

```
pkg_origin=acme
pkg_name=myapp

pkg_version() {
  cat "$SRC_PATH/version.txt"
}

do_before() {
  do_default_before
  update_pkg_version
}
```

Here is the same Plan which uses a Git command to compute the number of
commits in the repository as the version. Note that since we are using
the `git` command, we are adding the `core/git` package as a build
dependency:

```
pkg_origin=acme
pkg_name=myapp
pkg_build_deps=(core/git)

pkg_version() {
  git rev-list master --count
}

do_before() {
  do_default_before
  update_pkg_version
}
```

Finally, here is a partial implementation of a `cacerts` Plan which
greps a date (YYYY.mm.dd) from the CA certs file itself. The
`pkg_version()` function does some grep'ing and sed'ing before using the
date program to format the final version string to standard output. As
the downloaded file is required before running the version logic, the
`update_pkg_version()` helper is called in the `do_download()` build
phase:

```
pkg_origin=me
pkg_name=cacerts
pkg_source=http://curl.haxx.se/ca/cacert.pem

pkg_version() {
  local build_date

  # Extract the build date of the certificates file
  build_date=$(cat $HAB_CACHE_SRC_PATH/$pkg_filename \
    | grep 'Certificate data from Mozilla' \
    | sed 's/^## Certificate data from Mozilla as of: //')

  date --date="$build_date" "+%Y.%m.%d"
}

do_download() {
  do_default_download
  update_pkg_version
}

 # ...
```

Closes #2212

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>